bugfix:  - UnicodeEncodeError fixed when batch job is created  - cli arg for audio duration accepts float values

+++ src/cli/Main.py
@@ -113,7 +113,7 @@ def main():
     parser.add_option("-t", "--profile", help=profStr + " [default: %default]", default=3, type="int")
     parser.add_option("-n", "--videonorm", help="n=NTSC, p=PAL [default: %default]", default="p")
     parser.add_option("-f", "--format", help=formatStr + " [default: %default]", default=1, type="int")
-    parser.add_option("-l", "--length", help=_(u"total length of the PhotoFilmStrip (seconds)"), type="int", metavar="SECONDS")
+    parser.add_option("-l", "--length", help=_(u"total length of the PhotoFilmStrip (seconds)"), type="float", metavar="SECONDS")
     parser.add_option("-a", "--audio", help=_(u"use audiofile as audiotrack (use --length to limit the movie length)"), metavar="MP3")
     
     options = parser.parse_args()[0]
